feat: move modalContents to src/components/modalContent
