+++ src/pages/listPage/ListPage.jsx
@@ -3,11 +3,11 @@ import Modal from '@/components/Modal';
 import Header from '@/components/Header';
 import CreditSection from '@/pages/listPage/CreditSection';
 import { getCredits } from '@/utils/creditStorage';
-import RechargeModalContent from '@/pages/listPage/RechargeModalContent';
+import RechargeModalContent from '@/components/modalContent/RechargeModalContent';
 import CreditRechargeSuccess from '@/pages/listPage/CreditRechargeSuccess';
-import CreditShortageModalContent from '@/pages/listPage/CreditShortageModalContent';
+import CreditShortageModalContent from '@/components/modalContent/CreditShortageModalContent';
 import DonationsList from '@/pages/listPage/DonationsList';
-import DonationModalContent from '@/pages/listPage/DonationModalContent';
+import DonationModalContent from '@/components/modalContent/DonationModalContent';
 import leftTopGradient from '@/assets/images/leftTopGradient.png';
 import DonationSuccess from './DonationSuccess';
 import MonthlyChartSection from './MonthlyChartSection';
